Type: Adding the indications to download the data, data exploration and Logistic Regression estimation
First, indications to download the data were added. When the repo is cloned locally and lfs is configured, the dataset will be correctly downloaded. However, from sagemaker, the best option to obtain the dataset is through wget and unzip.

Second, the data exploration and visualization of features was  performed.

Third, two versions of  Logistic Regression were estimated: Accounting for class imbalance and not accounting for it.